+++ i18n/locales/en/common.json
@@ -186,14 +186,12 @@
   "Go to Dashboard": "Go to Dashboard",
   "Go to confirmation": "Go to confirmation",
   "Got it, thanks!": "Got it, thanks!",
-  "Got the Lisk Mobile App?": "Got the Lisk Mobile App?",
   "Height": "Height",
   "Height distribution": "Height distribution",
   "Help": "Help",
   "Help improve Lisk by allowing Lisk to gather anonymous usage data used for analytical purposes.": "Help improve Lisk by allowing Lisk to gather anonymous usage data used for analytical purposes.",
   "Hide": "Hide",
   "Hide balance and transactions amounts": "Hide balance and transactions amounts",
-  "Hide the QR code": "Hide the QR code",
   "High": "High",
   "How is Lisk transparent?": "How is Lisk transparent?",
   "How we recommend to store it.": "How we recommend to store it.",
@@ -359,7 +357,6 @@
   "Removed": "Removed",
   "Removed votes": "Removed votes",
   "Report the error via E-Mail": "Report the error via E-Mail",
-  "Request {{token}}": "Request {{token}}",
   "Requested amount": "Requested amount",
   "Required": "Required",
   "Restart now": "Restart now",
@@ -371,7 +368,6 @@
   "Save changes": "Save changes",
   "Save it on an encrypted hard drive: USB key or a backup drive": "Save it on an encrypted hard drive: USB key or a backup drive",
   "Save your passphrase": "Save your passphrase",
-  "Scan address": "Scan address",
   "Search by name or address": "Search by name or address",
   "Search within the network...": "Search within the network...",
   "Second passphrase": "Second passphrase",
@@ -393,9 +389,7 @@
   "Session timeout": "Your session was timed out after 10 minutes of no network activity. Please sign in to continue using your account.",
   "Settings": "Settings",
   "Settings saved!": "Settings saved!",
-  "Sharing link": "Sharing link",
   "Show": "Show",
-  "Show the QR code": "Show the QR code",
   "Sign Message": "Sign Message",
   "Sign a message": "Sign a message",
   "Sign in": "Sign in",
