+++ src/components/screens/request/request.css
@@ -1,4 +1,4 @@
-@import '../../../../../../app/mixins.css';
+@import '../../../app/mixins.css';
 
 .container {
   display: flex;
