+++ src/components/screens/request/requestBtc.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { AutoResizeTextarea } from '../../../../../toolbox/inputs';
+import { AutoResizeTextarea } from '../../toolbox/inputs';
 import RequestWrapper from './requestWrapper';
 import styles from './request.css';
