+++ src/components/screens/wallet/overview/accountInfo/index.js
@@ -47,15 +47,9 @@ const AccountInfo = ({
           />
         </div>
         <div className={styles.helperIcon}>
-          <Tooltip
-            tooltipClassName={styles.qrCodeWrapper}
-            className={styles.qrCode}
-            position="bottom"
-            title={t('Scan address')}
-            content={<Icon name="qrCodeActive" className={styles.qrCodeIcon} />}
-          >
-            <QRCode value={address} size={154} />
-          </Tooltip>
+          <DialogLink component="request">
+            <Icon name="qrCodeActive" className={styles.qrCodeIcon} />
+          </DialogLink>
         </div>
         {
           host !== address ? (
