+++ src/components/screens/wallet/overview/balanceInfo/balanceInfo.css
@@ -55,18 +55,6 @@
   justify-content: center;
   font-weight: normal;
 
-  & > div {
-    flex-grow: 1;
-  }
-
-  & > div:nth-child(1) {
-    padding-right: 5px;
-  }
-
-  & > div:nth-child(2) {
-    padding-left: 5px;
-  }
-
   & .button {
     flex-grow: 1;
 
@@ -84,15 +72,4 @@
       background-color: var(--color-btc);
     }
   }
-}
-
-.requestDropdown {
-  left: auto;
-  right: -90px;
-  transform-origin: calc(100% - 196px) -5px;
-}
-
-.requestButton {
-  width: 100%;
-  background-color: var(--color-white);
 }
